Return message to avoid a null crash.

+++ index.js
@@ -152,6 +152,8 @@ router.route('/movies')
         //Weakness: This will find the first instance and delete that.
         //First check if the movie even exists.
         Movie.findOne({ title: req.body.title}).select('title').exec(function(err, movie) {
+            if(movie === null)
+                return(res.status(409).send({success: false, msg: 'Movie not found.'}));
             try {
                 Movie.deleteOne({title: req.body.title});
             } catch(e) {
